+++ OnlineUIComponents/ChatLogManager.cs
